Full structures support
- trees generation
- custom PRNG for structures
- changed structures loading
- fixed tree schematic
- changed seed type

+++ Assets/Scripts/Game.cs
@@ -61,7 +61,7 @@ public class Game : MonoBehaviour
 {
     [NonSerialized] public static float TickRate = 20;
     [NonSerialized] public static int Level = 0;
-    [NonSerialized] public static long Seed;
+    [NonSerialized] public static int Seed;
 
     static class Tiles
     {
@@ -106,7 +106,7 @@ public class Game : MonoBehaviour
     void Start()
     {
         Random rand = new Random();
-        Seed = (long)rand.NextDouble();
+        Seed = (int)rand.NextDouble();
         
         // initialize static classes
         NoiseGen.Init();
